// frontend/smartqurylys/src/app/core/enums/activity-log.enum.ts
export enum ActivityActionType {
  REQUEST_ACCEPTANCE = 'REQUEST_ACCEPTANCE',
  ACCEPTED_ACCEPTANCE = 'ACCEPTED_ACCEPTANCE',
  REJECTED_ACCEPTANCE = 'REJECTED_ACCEPTANCE',
  PROJECT_UPDATED = 'PROJECT_UPDATED',
  STAGE_UPDATED = 'STAGE_UPDATED',
  STAGE_DELETED = 'STAGE_DELETED',
  FILE_ADDED = 'FILE_ADDED',
  FILE_DELETED = 'FILE_DELETED',
  DOCUMENT_ADDED = 'DOCUMENT_ADDED',
  DOCUMENT_DELETED = 'DOCUMENT_DELETED',
  PARTICIPANT_INVITED = 'PARTICIPANT_INVITED',
  PARTICIPANT_JOINED = 'PARTICIPANT_JOINED',
  PARTICIPANT_REMOVED = 'PARTICIPANT_REMOVED'
}

export enum ActivityEntityType {
  PROJECT = 'PROJECT',
  STAGE = 'STAGE',
  FILE = 'FILE',
  DOCUMENT = 'DOCUMENT',
  PARTICIPANT = 'PARTICIPANT'
}